OCR/ocr_requirements.txt
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.0.0
redis==5.0.1
openai==1.3.7
numpy==1.24.3